only restaurant owner can add/update/delete food-items to/from it

+++ controllers/foodItems.js
@@ -45,11 +45,27 @@ exports.getFoodItem = AsyncHandler(async (req, res, next) => {
 });
 
 //  @desc       Add new FoodItem
-//  @route      POST /api/v1/food-items
+//  @route      POST /api/v1/restaurants/:restaurantId/food-items/:id
 //  @access     Private
 exports.addFoodItem = AsyncHandler(async (req, res, next) => {
+  const restaurant = await Restraurant.findById(req.params.restaurantId);
+  if (!restaurant) {
+    return next(
+      new ErrorResponse(
+        `no Restaurant found with id: ${req.params.restaurantId}`,
+        400
+      )
+    );
+  }
   req.body.restaurant = req.params.restaurantId;
-
+  //  check if loged user is truly owner of restaurant or admin
+  if (
+    req.user.id !== restaurant.owner.toString() &&
+    req.user.role !== "admin"
+  ) {
+    return next(new ErrorResponse("not authorized to access resource", 403));
+  }
+  req.body.user = req.user._id;
   const item = await FoodItem.create(req.body);
   res.status(201).json({
     success: true,
@@ -64,6 +80,10 @@ exports.updateFoodItem = AsyncHandler(async (req, res, next) => {
   const food_item = await FoodItem.findById(req.params.id);
   if (!food_item) {
     return next(new ErrorResponse("no FoodItem found", 404));
+  }
+  //  check if loged user is truly owner of restaurant or admin
+  if (req.user.id !== food_item.user.toString() && req.user.role !== "admin") {
+    return next(new ErrorResponse("not authorized to access resource", 403));
   }
   const updated = await FoodItem.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
@@ -82,6 +102,10 @@ exports.deleteFoodItem = AsyncHandler(async (req, res, next) => {
   const food_item = await FoodItem.findById(req.params.id);
   if (!food_item) {
     return next(new ErrorResponse("no FoodItem found", 404));
+  }
+  //  check if loged user is truly owner of restaurant or admin
+  if (req.user.id !== food_item.user.toString() && req.user.role !== "admin") {
+    return next(new ErrorResponse("not authorized to access resource", 403));
   }
   await FoodItem.findByIdAndDelete(req.params.id);
   res.status(200).json({
